almost

+++ pythagorean-triplet/pythagorean_triplet.h
@@ -1,6 +1,6 @@
 #ifndef PYTHAGOREAN_TRIPLET_H
 #define PYTHAGOREAN_TRIPLET_H
-
+#include <stdint.h>
 typedef struct triplet_t
 {
     int a,b,c;
@@ -8,7 +8,7 @@ typedef struct triplet_t
 typedef struct triplets_t
 {
     triplet_t* triplets;
-    int count;
+    uint16_t count;
 }triplets_t;
 
 triplets_t* triplets_with_sum(int sum);
